Added the backend packages. Renamed the notification classes and put them in their right place.
